+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+TBD
+
+### Changed
+- API call: `getAlbums(withUPC: String...` changed to `getAlbums(withUPCs albumUPCs: [String]...`
+
+### Deprecated
+TBD
+
+### Removed
+TBD
+
+### Fixed
+TBD
+
+### Security
+TBD
+
 ## [0.2.0] - 2026-02-24
 
 ### Added